vulnerable-services/user-service/requirements.txt
# User Service Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
prometheus-client==0.19.0
pydantic==2.5.0

